+++ Assets/Source/Planet/PlanetState.cs
@@ -349,7 +349,7 @@ namespace Planet
         }
 
 
-        public ParticleEntity AddParticle(Vec2f position, Vec2f velocity, ParticleType type, float health = 1.0f)
+        public ParticleEntity AddParticle(Vec2f position, Vec2f velocity, ParticleType type, int health = 1)
         {
             Utils.Assert(ParticleList.Length < PlanetEntityLimits.ParticleLimit);
